Enable capability to handle tests that may cause a panic
If a test cause a panic, the json output from go test does not
show "Action: fail". In order to handle this, create a visit map,
and check if the test is runned but not record as fail, skip, and
pass. If such test exist, record it as failing test.